Interim commit of network reorg.
